Refactor webapp into package and add optional dependencies
